+++ example.py
@@ -1,7 +1,7 @@
 from datetime import timedelta
 
 from py_image_dedup.config.deduplicator_config import DeduplicatorConfig
-from py_image_dedup.library.image_match_deduplicator import ImageMatchDeduplicator
+from py_image_dedup.library.deduplicator import ImageMatchDeduplicator
 
 config = DeduplicatorConfig()
 # config.ELASTICSEARCH_HOST.value = "192.168.2.24"
